requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
AutoDict==2.0.0
av==10.0.0
beautifulsoup4==4.12.1
bitarray==2.3.4
blis==0.7.7
Bottleneck==1.3.7
catalogue==2.0.7
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
chumpy==0.70
click==8.1.3
cmake==3.26.1
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.6
dataclasses==0.6
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.6.0
eq==0.0.10
fairseq==0.10.2
fastai==1.0.55
fastprogress==1.0.3
filelock==3.10.0
filterpy==1.4.5
flatten-dict==0.2.0
fonttools==4.39.2
freetype-py==2.3.0
FrEIA==0.2
frozenlist==1.3.3
fsspec==2023.3.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.3
hydra-core==1.0.7
idna==3.4
imageio==2.26.1
imgaug==0.4.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jpeg4py==0.1.4
jsonargparse==4.20.0
kiwisolver==1.4.4
kornia==0.6.9
langcodes==3.3.0
lazy_loader==0.2
lightning-utilities==0.8.0
lit==16.0.0
littleutils==0.2.2
llvmlite==0.37.0
loguru==0.3.2
MarkupSafe==2.1.2
matplotlib==3.5.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.7
networkx==3.0
numba==0.54.1
numexpr==2.8.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ogb==1.3.5
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
outdated==0.2.2
packaging==23.0
pandas==1.5.3
pathlib2==2.3.7.post1
pathtools==0.1.2
pathy==0.6.1
Pillow==9.4.0
portalocker==2.3.2
preshed==3.0.6
protobuf==4.22.1
psutil==5.9.4
pydantic==1.8.2
pyDeprecate==0.3.2
pyglet==2.0.5
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.36
python-dateutil==2.8.2
pytorch-lightning==1.9.4
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2021.10.8
requests==2.28.2
sacrebleu==2.0.0
sacremoses==0.0.46
scikit-image==0.19.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.17.0
setproctitle==1.3.2
setuptools-scm==7.1.0
shapely==2.0.1
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
smplx==0.1.28
soupsieve==2.4
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
subword-nmt==0.3.7
sympy==1.11.1
thinc==8.0.15
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.12
tomli==2.0.1
torch==2.0.0
torchaudio==0.9.1
torchgeometry==0.1.2
torchmetrics==0.10.3
torchvision==0.15.1
tqdm==4.65.0
trimesh==3.8.12
triton==2.0.0
typer==0.4.1
typeshed-client==2.2.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.14.0
wasabi==0.9.1
wilds==2.0.0
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0
